Example: always print the base addresses as longs.

+++ example.c
@@ -25,7 +25,7 @@ int main(void)
       c = pci_read_byte(dev, PCI_INTERRUPT_PIN);				/* Read config register directly */
       printf("%02x:%02x.%d vendor=%04x device=%04x class=%04x irq=%d (pin %d) base0=%lx\n",
 	     dev->bus, dev->dev, dev->func, dev->vendor_id, dev->device_id,
-	     dev->device_class, dev->irq, c, dev->base_addr[0]);
+	     dev->device_class, dev->irq, c, (long) dev->base_addr[0]);
     }
   pci_cleanup(pacc);		/* Close everything */
   return 0;
